Update AVR_Miner.py
Changing the socket .send() to .sendall() in lines #800 and #946. I have found that sometimes, depending on the network speed or latency, the message from the server can take some time to arrive or can arrive fragmented, this can cause some corruption on the code that could be dragged for the next jobs and never recover from it. sendall() seems to be more stable. On my side is working good but it would be better if someone else can test it.