Update AVR_Miner.py
It's way better to use the "select" package for handling the messages sent by the server. Using this we don't have to worry on waiting for the buffer to be ready, select does it for you. Using just plain recv for reading the buffer is not enough, since recv doesn't wait for nobody and can return an empty buffer if the server hasn't finish sending the message or if the client has latency or slow connection. I've added some global variables inside functions to ensure we are getting the correct value.