Update ESP32_Code.ino
Adding a semaphore for handling the shared resources (esp_sha), removing the mbed implementation for SHA1 it's not needed anymore, adding a flush after every buffer reading and before all the buffer writing to ensure that there is no corrupted data on the buffer. Also a minor change on lines 96 and 204, there is a String conversion of a String variable, is not necessary and it is under the main loop for the job calculation so with this change there is an speed improvement. Finally I removed the yield command on the job calculation loop since I think is not necessary anymore and it also improves a the speed a little bit. I have tested the code successfully for a couple of hours, I will appreciate if someone else can test it too.