Update ESP32_Code.ino
- Move the declarations of variables and memory allocations out of the loops so they're done just once.
- Implementing the tasks as infinite loops, there is no need to kill the tasks at any error occurrence.
- Adding some memset calls to clear the buffers before any job calculation.
- Changing the way to connect to the server, if the client doesn't connect to the server return to the beginning of the loop and try again. After connecting wait for the server version before asking for a job, this will prevent the mixing of server version and job messages.
- Changing the socket timeout to 1 second, 2 seconds is not necessary at least with the tests I have run.

Since a lot of the variable declarations are moved the changes look messy, I didn't wanna to requested a change like this but I think there is no other way to change it.
By the way, with these changes the hashrate will double or be close to that so maybe the difficulty rules will need to be changed. Please someone test this code and see if it's working better.