Update ESP32_Code.ino
Slight modification for making the OTA implementation optional. This is for some boards like the ESP32-CAM that doesn't support OTA directly (some people say that it is possible but you need to change the core files), the user just need to comment line 36 to disable OTA. This should be temporary until the Espressif team make the OTA possible for every board.