Update README.md
I've run some tests with the ESP32 as close as possible to my router and with a USB current monitor. I've found that the the hashrate goes between 21 and 26 KH/s it could get rarely to 27 but it is not common and mostly it will stay in 23, the daily profit goes between 8 and above 9 but never 10 and the current goes between 90mA to 160mA so in rare cases it could go up to 200mA. So with this results I think it's better to say that the ESP32 has an average hashrate of 23 KH/s, daily profit of 8-9 DUCO and average power 1 W (it's really more like peak power but in this case the worst scenario is better).