Update hash.c
Small change. This loop doesn't get to 8 because the maximum size of state.words is (SHA1_HASH_LEN / 4) and SHA1_HASH_LEN is 20. It's better to use the constant name instead of a fixed value. This has been pointed out by the compiler when you put the higher warning level.