Use of PROGMEM and make the code more generic
While I was working with the megaAVR architectures I found out that there were some portions of code that depend on the AVR architecture. Since the megaAVR is not strictly AVR architecture I had to change that but while I was looking at it and doing some research I found out that the use of PROGMEM and all the flash or program memory instructions are not exclusive to the AVR architectures, all the other architectures do what's necessary to handle them:

https://www.arduino.cc/reference/en/language/variables/utilities/progmem/
https://arduino.stackexchange.com/questions/65273/how-to-use-progmem-in-a-h-cpp-file-instead-of-a-ino-file

The use of #include "Arduino.h" is enough for this and is more generic. I have tested the code with the Teensy 4.1, a SAMD21 board, the Leonardo board and the ESP32 (as serial miner). I have checked and it also compile on the Raspberry Pi Pico, the megaAVR boards and STM32 but haven't tested them yet since I don't have any.