Optimize ESP8266 miner
This commit changes how the sequence of strings representing consecutive
integers is generated for the hasher.  Instead of utilizing the Arduino
String conversions, a Counter class is introduced, which can generate
consecutive integers much faster.

On my Wemos D1 mini this gives a hash rate increase from around 37 kH/s
to around 53 kH/s, which is a 43% improvement.