Addet Telnet & update shares without reload page
If the ESP8266 is not connected to the PC there is no possibility of output via the serial port. I added a Telnet server to the code. Now a connection can be established via the IP using a Telnet client such as Putty and the mining progress can be displayed. Secondly, I expanded the code so that the current shares, in addition to the hashrate, are updated via push without the page having to be reloaded.